bypass the syscall requests that has been serviced
there will be one syscall_main_t that proxy
syscalls for the leader simulator, and when
syscall_main_t is servicing a request, every HTIF
transactions involved will be recorded.

the recorded HTIF transactions will be copied to
each registered syscall_mirror_t, which proxy the
syscall requests for the trailer simulator.

the trailer simulator is all about repeating what
leader simulator did, so a syscall_mirror_t can
bypass the requests from a trailer simulator by
replaying the responding sequence received from
syscall_main_t.

in case of 721sim, this is to ensure each syscall
will be executed only once in spite of the
existance of two simulators (ISA and MICRO), and
another benifit is that the two simulator will
get their syscall request serviced in an exact
same way.