added ability to record & replay FESVR dev traffic
the recording is saved to a packet-based binary format. recordings will
be saved into a folder, and each core will have a dedicated recording
file in the folder. also, the folder contains a file that keeps the
expected target specification (memory size, core count, SHA256 of the
loaded ELF).

this recorded traffic can later be replayed to support target system
(for example service syscall request) without doing actual device
emulation. the main benefit of the replay mode is that the syscalls then
no longer need be submitted to the host OS, which allows to make a RISCV
workload really self-contained.

in order to use a pre-recorded FESVR device traffic to run a RISCV
workload, the target specification must fully match the expectation.
the traffic of each core will be replayed independently.

the flag for recording traffic is +dev-traffic-record=<output-folder>

the flag for replaying traffic is +dev-traffic-replay=<input-folder>